feat(l1-migration): move replay tool into movement-full-node
