--- l1-migration/replay/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "replay-blocks"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-repository.workspace = true
-homepage.workspace = true
-publish.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-aptos-types = { workspace = true }
-async-stream = { workspace = true }
-anyhow = { workspace = true }
-bcs = { workspace = true }
-clap = { workspace = true }
-futures = { workspace = true }
-movement-da-sequencer-client = { workspace = true }
-movement-da-sequencer-proto = { workspace = true }
-movement-types = { workspace = true }
-rocksdb = { workspace = true }
-serde_json = { workspace = true }
-tokio = { workspace = true }
-tokio-stream = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-aptos-rest-client = { workspace = true }
-
-[lints]
-workspace = true
