+++ networks/movement/movement-full-node/src/admin/l1_migration/replay/da_height.rs
@@ -1,8 +1,8 @@
-use crate::types::da::DaSequencerDb;
+use crate::admin::l1_migration::replay::types::da::DaSequencerDb;
 use clap::Parser;
 use std::path::PathBuf;
 
-#[derive(Parser)]
+#[derive(Parser, Debug)]
 #[clap(name = "da-height", about = "Extract synced block height from the DA-sequencer database")]
 pub struct DaHeight {
 	#[arg(help = "Path to the DA-Sequencer database")]
@@ -10,8 +10,8 @@ pub struct DaHeight {
 }
 
 impl DaHeight {
-	pub fn run(self) -> anyhow::Result<()> {
-		let db = DaSequencerDb::open(self.path)?;
+	pub fn run(&self) -> anyhow::Result<()> {
+		let db = DaSequencerDb::open(&self.path)?;
 		let height = db.get_synced_height()?;
 		println!("{}", height);
 		Ok(())
