+++ networks/movement/movement-full-node/src/admin/l1_migration/replay/mod.rs
@@ -1,12 +1,12 @@
-use crate::da_height::DaHeight;
-use crate::replay::DaReplayTransactions;
+use crate::admin::l1_migration::replay::da_height::DaHeight;
+use crate::admin::l1_migration::replay::replay::DaReplayTransactions;
 use clap::Parser;
 
 mod da_height;
 mod replay;
 mod types;
 
-#[derive(Parser)]
+#[derive(Parser, Debug)]
 #[clap(name = "Movement Da-Sequencer replay tool", author, disable_version_flag = true)]
 pub enum ApiReplayTool {
 	Replay(DaReplayTransactions),
@@ -14,7 +14,7 @@ pub enum ApiReplayTool {
 }
 
 impl ApiReplayTool {
-	pub async fn run(self) -> anyhow::Result<()> {
+	pub async fn execute(&self) -> anyhow::Result<()> {
 		match self {
 			ApiReplayTool::Replay(cmd) => cmd.run().await,
 			ApiReplayTool::ExtractDaHeight(cmd) => cmd.run(),
