+++ networks/movement/movement-full-node/src/node/mod.rs
@@ -1,4 +1,4 @@
-mod da_db;
+pub mod da_db;
 pub mod manager;
 pub mod partial;
 mod tasks;
